+++ inst/app/tools/analysis/sample_size_comp_ui.R
@@ -2,7 +2,9 @@
 # Sample size
 ###############################
 ssc_type <- c("Mean" = "mean", "Proportion" = "proportion")
-ssc_alternative <- c("Two sided" = "two.sided", "One sided" = "one.sided")
+# ssc_alternative <- c("Two sided" = "two.sided", "One sided" = "one.sided")
+# ssc_alternative <- c("Two sided" = "two.sided", "Less than" = "less", "Greater than" = "greater")
+ssc_alternative <- c("Two sided" = "two.sided", "Group 1 less than Group 2" = "less", "Group 1 greater than Group 2" = "greater")
 
 ## list of function arguments
 ssc_args <- as.list(formals(sample_size_comp))
@@ -23,8 +25,12 @@ output$ui_sample_size_comp <- renderUI({
         selected = state_init("ssc_type", "mean"), inline = TRUE
       ),
       numericInput(
-        "ssc_n", "Sample size (n):", min = 1,
-        value = state_init("ssc_n", NA), step = 1
+        "ssc_n1", "Sample size (n1):", min = 1,
+        value = state_init("ssc_n1", NA), step = 1
+      ),
+      numericInput(
+        "ssc_n2", "Sample size (n2):", min = 1,
+        value = state_init("ssc_n2", NA), step = 1
       ),
       conditionalPanel(
         condition = "input.ssc_type == 'mean'",
@@ -56,14 +62,13 @@ output$ui_sample_size_comp <- renderUI({
         "ssc_power", "Power:", min = 0, max = 1,
         value = state_init("ssc_power", 0.8), step = .05
       ),
-      numericInput(
-        "ssc_ratio", "Ratio (n1 / n2):", min = 0,
-        value = state_init("ssc_ratio", 1), step = .05
+      selectInput(
+        inputId = "ssc_alternative", label = "Alternative hypothesis:",
+        choices = ssc_alternative,
+        selected = state_single("ssc_alternative", ssc_alternative, "two.sided")
+
       ),
-      radioButtons(
-        inputId = "ssc_alternative", label = NULL, choices = ssc_alternative,
-        selected = state_init("ssc_alternative", "two.sided"), inline = TRUE
-      )
+      checkboxInput("ssc_show_plot", "Show plot" , state_init("ssc_show_plot", FALSE))
     ),
     help_and_report(
       modal_title = "Sample size (compare)", fun_name = "sample_size_comp",
@@ -72,12 +77,28 @@ output$ui_sample_size_comp <- renderUI({
   )
 })
 
+ssc_plot_width <- function() 650
+ssc_plot_height <- function() 650
+
 output$sample_size_comp <- renderUI({
   register_print_output("summary_sample_size_comp", ".summary_sample_size_comp")
+  register_plot_output(
+    "plot_sample_size_comp", ".plot_sample_size_comp",
+    width_fun = "ssc_plot_width",
+    height_fun = "ssc_plot_height"
+  )
 
   ## one output with components stacked
   ssc_output_panels <- tagList(
-    tabPanel("Summary", verbatimTextOutput("summary_sample_size_comp"))
+    tabPanel("Summary", verbatimTextOutput("summary_sample_size_comp")),
+    tabPanel(
+      "Plot",
+      conditionalPanel(
+        "input.ssc_show_plot == true",
+        download_link("dlp_ssc"),
+        plotOutput("plot_sample_size_comp", height = "100%")
+      )
+    )
   )
 
   stat_tab_panel(
@@ -98,6 +119,11 @@ output$sample_size_comp <- renderUI({
   summary(.sample_size_comp())
 })
 
+.plot_sample_size_comp <- reactive({
+  req(input$ssc_show_plot == TRUE)
+  plot(.sample_size_comp())
+})
+
 observeEvent(input$sample_size_comp_report, {
   ssc <- ssc_inputs()
   if (input$ssc_type == "mean") {
@@ -106,9 +132,33 @@ observeEvent(input$sample_size_comp_report, {
     ssc$delta <- ssc$sd <- NULL
   }
 
-  inp_main <- clean_args(ssc, ssc_args)
+  inp_out <- list("", "")
+  outputs <- "summary"
+  figs <- FALSE
+  if (isTRUE(input$ssc_show_plot)) {
+    inp_out[[2]] <- list(custom = FALSE)
+    outputs <- c("summary", "plot")
+    figs <- TRUE
+  }
+
   update_report(
-    inp_main = inp_main,
-    fun_name = "sample_size_comp", outputs = "summary", figs = FALSE
+    inp_main = clean_args(ssc, ssc_args),
+    fun_name = "sample_size_comp",
+    inp_out = inp_out,
+    outputs = outputs,
+    figs = figs,
+    fig.width = ssc_plot_width(),
+    fig.height = ssc_plot_height()
   )
 })
+
+download_handler(
+  id = "dlp_ssc",
+  fun = download_handler_plot,
+  fn = function() paste0("sample_size_comp_", input$ssc_type),
+  type = "png",
+  caption = "Save sample size comparison plot",
+  plot = .plot_sample_size_comp,
+  width = ssc_plot_width,
+  height = ssc_plot_height
+)