+++ NEWS.md
@@ -1,5 +1,6 @@
-# radiant.design 0.9.7.2
+# radiant.design 0.9.7.3
 
+* Use the `pwr` package for sample size calculations when comparing groups (i.e., Design > Sample size (compare))
 * Load a state file on startup by providing a (relative) file path or a url
 
 # radiant.design 0.9.7.0